implemented player action history and history checking, added new goal
now after performing any action, it will be recorded in the 'history'
library. When an action that would normally result in a stat change or
something else that should be non-repeatable, the game can now check the
character's history and return different consequences depending on the
number of times the action has been done before.