implemented objective system
new consequence flag for changing the player's goals. New goals are put
into the 'active_goals' variable. When a goal is completed, it is
removed from 'active_goals', and entered into 'completed_goals', either
marked 'completed' or 'failed'. Additional consequences can be triggered
upon initial receipt of the goal, or when it is completed or failed.