moved system command handling to its own function for readability
also found a bug in the goal change logic. I was able to complete the
goal without it being in progress first. However, it did not show up
until I performed the interaction that would add it in the first place,
resulting in the goal being listed as 'completed' in my active goals and
not in my finished goals.